New WMC parsing library: fixes namespace issues. Coords are now stored as double instead of float. Fixed a bug in creating WMC: bbox was taken from server, but it has to be provided by the client. Fixed standalone IM.
git-svn-id: https://geonetwork.svn.sourceforge.net/svnroot/geonetwork/branches/2.2.x@1301 ff403467-3f20-0410-99e5-ff581e626a08

+++ src/org/wfp/vam/intermap/kernel/Geonet.java
@@ -39,8 +39,8 @@ import org.apache.commons.httpclient.methods.*;
 import org.apache.commons.httpclient.cookie.CookiePolicy;
 
 public class Geonet {
-	
-	public static Element getGeonetRecords(float minx, float miny, float maxx, float maxy, int from, int to)
+
+	public static Element getGeonetRecords(double minx, double miny, double maxx, double maxy, int from, int to)
 		throws MalformedURLException, IOException, Exception
 	{
 		// Get initial state object
@@ -48,29 +48,29 @@ public class Geonet {
 		httpclient.getParams().setCookiePolicy(CookiePolicy.RFC_2109);
 		//HostConfiguration hConf = httpclient.getHostConfiguration(); // DEBUG
 		//hConf.setProxy("10.11.40.110", 8080); // DEBUG
-		
+
 		// portal.search request
 		String request = "http://www.fao.org/geonetwork/srv/en/portal.search?extended=on&remote=off&region=0000&selregion=%3B180%3B-180%3B-90%3B90&relation=overlaps&any=&title=&abstract=&themekey=&radfrom=&siteId=&category=&hitsPerPage=10"
 			+ "&northBL=" + maxy + "&westBL=" + minx + "&eastBL=" + maxx + "&southBL=" + miny;
 		request += "&from=&to=";
-		
+
 		System.out.println("request: " + request);
-		
+
 		HttpMethod method = new GetMethod(request);
-		
+
 		int statusCode = httpclient.executeMethod(method);
 		if (statusCode == HttpStatus.SC_OK)
 			method.releaseConnection();
 		else
 			throw new Exception();
-		
+
 		// portal.present request
 		request = "http://www.fao.org/geonetwork/srv/en/portal.present?siteId=&hitsPerPage=10";
 		if (from != -1 && to != -1)
 			request += "&from=" +  from + "&to=" + to;
 		else
 			request += "&from=&to=";
-		
+
 		method = new GetMethod(request);
 		statusCode = httpclient.executeMethod(method);
 		if (statusCode == HttpStatus.SC_OK)
@@ -82,8 +82,8 @@ public class Geonet {
 		}
 		else
 			throw new Exception();
-		
+
 	}
-	
+
 }
